:recycle: simplify responsive css definitions
I thought it might be good to approach css from a mobile
first approach. Meaning that when we write style for various
components we can write them with mobile in mind. With mobile
as the default styling for components we can then add break
points for any larger sizes.

Adding this as a separate commit so that it can be reverted
if anyone sees issues with this change.