Move HL 3.0 license files into own folder _build-dependency-hl-3.0/
The reasoning for this is that with the Netlify function
build process and the license builder functionality overall
the way the HL 3.0 license is structured is so different from
the other licenses under version that I thought it would be
good to reorganize.

This does not interfere with the resulting URL structure. Due
to the Netlify proxy users will still be able to use /version/3/0/full.html
to reach the HL 3.0 license.