:zap::recycle: skip installing devDependencies on deploy
This change installs fewer packages in production. Also,
we instruct Husky to skip installing its git hooks.

Moved husky to dependencies because it's only 6kb and needs
to exist so that husky install kan be successfully skipped.

Moved AVA to dependencies because it's required for the
build step to succeed.